+++ packages/spry/CHANGELOG.md
@@ -16,8 +16,9 @@ dependencies:
 ## What's Changed
 
 1. **Docs**: fix `route.path` to `route.segments`, Thanks [@utamori](https://github.com/utamori)
-2. **Feature**: Support `RethrowEception`, Catch exception to next filter process.
-3. **Feature**: `request.json` supports implicit type conversion.
+2. **Feature**: Support `RethrowEception`, Catch exception to next filter process
+3. **Feature**: `request.json` supports implicit type conversion
+4. **Feature**: Added built-in `Blob` export
 
 ## Credits